Export Test - Do Not Merge
Move matchers' tests from *generated* to common files.

PiperOrigin-RevId: 320954068
